Add files via upload
WebsiteRandomWatermark is a powerful JavaScript framework designed specifically for adding screen watermarks to websites. It allows users to customize the number of watermarks, randomly set the position of the watermark, and add animated effects to the watermark. In addition, the framework also provides functionality to prevent users from copying webpage content, as well as the option to pop up a prompt window when users attempt to copy content.